modebar: make module more consistent with other components
by:
- making modebar/index.js require what needs to be exposed
  for other parts of the code
- in consequence old modebar/index.js -> modebar/modebar.js
- var ModeBar = require('modebar/') and use ModeBar.manage
  in plot_api.js
