+++ test/jasmine/tests/modebar_test.js
@@ -1,6 +1,6 @@
 var d3 = require('d3');
 
-var createModeBar = require('@src/components/modebar');
+var createModeBar = require('@src/components/modebar/modebar');
 var manageModeBar = require('@src/components/modebar/manage');
 
 var Plotly = require('@lib/index');
